ml/requirements.txt
torch==2.4.0
torchvision==0.19.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
matplotlib==3.9.2
pillow==10.4.0
pyyaml==6.0.2
